+++ xenon/repl/repl.py
@@ -139,7 +139,11 @@ class REPL:
         from xenon.repl.model_pool import ModelPool
         from xenon.repl.auto_router import AutoRouter
         self.model_pool = ModelPool()
-        self.auto_router = AutoRouter(self.model_pool, context_manager=self.ctx_mgr)
+        self.auto_router = AutoRouter(
+            self.model_pool,
+            context_manager=self.ctx_mgr,
+            cache_tracker=self._cache_tracker,
+        )
         self.status_bar._auto_router = self.auto_router  # for "auto" display
 
         # 会话状态，供命令处理器共享